Wave 19 deep-crack: player/intersect/sky/objseq register cracks
player fn_802AB1D0 91->100 (opt_level 2 index form), fn_80296D20 97->100 (#130
re-derive state), objAnimFn shared-return merge, #62 (int)-launder transform args;
intersect 67->73 at 100% (fn_8007D670/C664/CAF4/screenImageDraw + real GXSetTevIndirect
arg bugs); sky2_update 100 (#130), snowFreeSnowCloud 100 (#17), ObjSeq_UpdateCurvePosition
100 (#45 FP decl-order), playerTailFn/videoSwapFrameBuffers 100, fn_80065768 95->97.7.
Per-file 3-way merge onto origin (player convergent w/ origin).

+++ src/main/pi_dolphin.c
@@ -7499,13 +7499,13 @@ extern void* displayFrameBuffer;
 #pragma peephole on
 void videoSwapFrameBuffers(void)
 {
-    u16 sync;
+    int sync;
     int tok[3];
     char fifo[140];
 
     lbl_803DCCA0 = lbl_803DCCA0 + 1;
     sync = GXReadDrawSync();
-    if (sync == (u16)(lbl_803DCCAA + 1))
+    if ((u16)sync == (u16)(lbl_803DCCAA + 1))
     {
         lbl_803DCCAA = sync;
         if (displayFrameBuffer == externalFrameBuffer0)
